backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Cache
redis[hiredis]==5.1.1

# Browser Automation
playwright==1.47.0

# AI
google-generativeai==0.8.3

# Vector DB
chromadb==0.5.15
sentence-transformers==3.1.1

# HTTP
httpx==0.27.2

# Logging
structlog==24.4.0
colorama==0.4.6

# Observability
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0

# Utils
python-dotenv==1.0.1
tenacity==9.0.0
